refactor: extract SpendingTrendlineChart component
